Added the check for the project icon as we don't want to delete this for the Project Update

+++ Sources/App/Controllers/AssetController.swift
@@ -100,7 +100,9 @@ final class AssetController {
     }
     
     for asset in assets {
-      try asset.delete()
+      if !asset.project_icon {
+        try asset.delete()
+      }
     }
     
     return try JSON(node: ["message": "Assets deleted!"])
